Remove ESBuild file, Renaming data to constants
